chore: split up gatling plugin settings
Motivation:
Too much configuration inside `GatlingPlugin`

Modification:
* Add settings package with all previous settings split up
* Reduce comments sizes
